+++ src/main/java/com/upc/backendnutrimiski/models/api/ApiDishesRequest.java
@@ -12,7 +12,7 @@ import java.util.List;
 public class ApiDishesRequest {
 
         Integer age;
-        Double weight;
+        Integer weight;
         Integer height;
         String activity;
         String sex;
